fix pr comments
