+++ src/panels/GettingStartedPanel/GettingStartedPanel.tsx
@@ -2,9 +2,7 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button, Group, Stack } from '@mantine/core';
 
-import { useOpenSpaceApi } from '@/api/hooks';
 import { Layout } from '@/components/Layout/Layout';
-import { usePropertyOwnerVisibility } from '@/hooks/propertyOwner';
 import { ArrowLeftIcon, ArrowRightIcon } from '@/icons/icons';
 import { useWindowLayoutProvider } from '@/windowmanagement/WindowLayout/hooks';
 
@@ -22,10 +20,6 @@ export function GettingStartedPanel() {
   const navigationSteps = useNavigationSteps();
   const timeSteps = useTimeSteps();
   const contentSteps = useContentSteps();
-  const luaApi = useOpenSpaceApi();
-  const { setVisibility: setVisibleEsriViirsCombo } = usePropertyOwnerVisibility(
-    'Scene.Earth.Renderable.Layers.ColorLayers.ESRI_VIIRS_Combo'
-  );
 
   const { t } = useTranslation('panel-gettingstartedtour');
 
@@ -63,37 +57,12 @@ export function GettingStartedPanel() {
     if (isLastStep) {
       closeWindow('gettingStartedTour');
     } else {
-      if (isLastIntroductionStep) {
-        setupGettingStartedTour();
-      }
       setStep(step + 1);
     }
   }
 
   function onClickPrev() {
     setStep(step - 1);
-  }
-
-  function setupGettingStartedTour(): void {
-    luaApi?.setPropertyValueSingle('RenderEngine.BlackoutFactor', 0.0, 1.0);
-    setTimeout(() => {
-      luaApi?.navigation.jumpTo('Earth');
-      luaApi?.action.triggerAction('os.earth_standard_illumination');
-      luaApi?.setPropertyValue(
-        'Scene.Earth.Renderable.Layers.ColorLayers.*.Enabled',
-        false
-      );
-      setVisibleEsriViirsCombo(true);
-      luaApi?.time.setPause(false);
-      luaApi?.time.interpolateDeltaTime(1);
-      luaApi?.fadeIn('Scene.EarthTrail.Renderable');
-      luaApi?.setPropertyValue(
-        'Scene.Mars.Renderable.Layers.ColorLayers.CTX_Mosaic_Sweden.Enabled',
-        false
-      );
-      luaApi?.fadeOut('Scene.Constellations.Renderable');
-      luaApi?.setPropertyValueSingle('RenderEngine.BlackoutFactor', 1.0, 1.0);
-    }, 1000);
   }
 
   return (
